port {urltools}	R Documentation

Get or set a URL's port

Description

as in the lubridate package, individual components of a URL can be both extracted or set using the relevant function call - see the examples.

Usage

port(x)

port(x) <- value

Arguments

x
a URL, or vector of URLs

value
a replacement value (or vector of replacement values) for x's port. If NULL, the port will be entirely removed.

See Also

scheme, domain, path, parameters and fragment for other accessors.

Examples

# Get the port
example_url <- "http://cran.r-project.org:80/submit.html"
port(example_url)

# Set the port
port(example_url) <- "12"

# Remove the port
port(example_url) <- NULL

[Package urltools version 1.7.3.1 Index]