suffix_dataset {urltools}	R Documentation

Dataset of public suffixes

Description

This dataset contains a registry of public suffixes, as retrieved from and defined by the public suffix list. It is sorted by how many periods(".") appear in the suffix, to optimise it for suffix_extract. It is a data.frame with two columns, the first is the list of suffixes and the second is our best guess at the comment or owner associated with the particular suffix.

Usage

data(suffix_dataset)

Format

A data.frame of 8030 rows and 2 columns

Note

Last updated 2016-07-31.

See Also

suffix_extract for extracting suffixes from domain names, and suffix_refresh for getting a new, totally-up-to-date dataset version.

[Package urltools version 1.7.3.1 Index]