suffix_refresh {urltools}	R Documentation

Retrieve a public suffix dataset

Description

urltools comes with an inbuilt dataset of public suffixes, suffix_dataset. This is used in suffix_extract to identify the top-level domain within a particular domain name.

While updates to the dataset will be included in each new package release, there's going to be a gap between changes to the suffixes list and changes to the package. Accordingly, the package also includes suffix_refresh, which generates and returns a fresh version of the dataset. This can then be passed through to suffix_extract.

Usage

suffix_refresh()

Value

a dataset equivalent in format to suffix_dataset.

See Also

suffix_extract to extract suffixes from domain names, or suffix_dataset for the inbuilt, default version of the data.

Examples

## Not run: 
new_suffixes <- suffix_refresh()

## End(Not run)

[Package urltools version 1.7.3.1 Index]