tld_refresh {urltools}	R Documentation

Retrieve a TLD dataset

Description

urltools comes with an inbuilt dataset of top level domains (TLDs), tld_dataset. This is used in tld_extract to identify the top-level domain within a particular domain name.

While updates to the dataset will be included in each new package release, there's going to be a gap between changes to TLDs and changes to the package. Accordingly, the package also includes tld_refresh, which generates and returns a fresh version of the dataset. This can then be passed through to tld_extract.

Usage

tld_refresh()

Value

a dataset equivalent in format to tld_dataset.

See Also

tld_extract to extract suffixes from domain names, or tld_dataset for the inbuilt, default version of the data.

Examples

## Not run: 
new_tlds <- tld_refresh()

## End(Not run)

[Package urltools version 1.7.3.1 Index]