url_compose {urltools}	R Documentation

Recompose Parsed URLs

Description

Sometimes you want to take a vector of URLs, parse them, perform some operations and then rebuild them. url_compose takes a data.frame produced by url_parse and rebuilds it into a vector of full URLs (or: URLs as full as the vector initially thrown into url_parse).

This is currently a 'beta' feature; please do report bugs if you find them.

Usage

url_compose(parsed_urls)

Arguments

parsed_urls
a data.frame sourced from url_parse

See Also

scheme and other accessors, which you may want to run URLs through before composing them to modify individual values.

Examples

#Parse a URL and compose it
url <- "http://en.wikipedia.org"
url_compose(url_parse(url))

[Package urltools version 1.7.3.1 Index]