urltools {urltools}	R Documentation

Tools for handling URLs

Description

This package provides functions for URL encoding and decoding, parsing, and parameter extraction, designed to be both fast and entirely vectorised. It is intended to be useful for people dealing with web-related datasets, such as server-side logs.

See Also

the package vignette.

[Package urltools version 1.7.3.1 Index]